% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supply_save.R
\name{supply_save}
\alias{supply_save}
\title{Save Talent Supply validated files}
\usage{
supply_save(country_name, year)
}
\arguments{
\item{country_name}{Country for which the saves are going to be saved}

\item{year}{Year for which the saves are going to be saved}
}
\description{
This function allows you to save validated Talent Supply files ready to be uploaded to data base.
}
\examples{
save_validated_files("South_Africa",2017)
}
\keyword{save}
