% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supply_validation.R
\name{supply_validation}
\alias{supply_validation}
\title{Talent Supply Validation Function}
\usage{
supply_validation(year, country_name, user, password, occupation_type_id,
  level_occupation, Are_Onet_files_available, host)
}
\arguments{
\item{year}{Year for which we are validating the data}

\item{country_name}{Country for which we are validating the data.E.g, "South_Africa"}

\item{user}{Data Base user name E.g, "user"}

\item{password}{Data base password E.g, "password"}

\item{occupation_type_id}{Country Occupation type id as in data base}

\item{level_occupation}{Level of occupation that we are validating, this number correspond to the column level as in table norm_prod.occupations}

\item{Are_Onet_files_available}{"yes" if the files with Onet supply are available and ready to be validated}

\item{host}{Indicate Hostname server}
}
\description{
This function allows you to validate the talent supply counts by Local and Onet taxonomies.This files must be saved in the data science Talent Supply
directory in share drive. The files with local occupation taxonomy supply counts must be saved in the Input_files folder.
The files with Onet occupation taxonomy supply counts must be saved in the Output_files folder.
A report with local occupation taxonomy is created for each available year in order to facilitate to the Quantitative Analyst the visual analysis
of the data sets if necessary.
The validation script executes the following validations, in the order specified below:
Local codes files validation
\enumerate{
\item Verify that the number of available local codes in the file at Country level is equal to the number of codes available in the State level file
\item Verify that the sum of total supply crossed local codes at State level is equal to Country level
\item Verify that the sum of supply of each occupation at Country level is less or equal than its State level counterpart. If test #3 fails, a file with all codes for which this test does not respect the constraint is saved in the error folder
\item Verify that the sum by occupation at MSA level is equal or less than its State counterpart. If test #4 fails, a file with all codes for which this test does not respect the constraint is saved in the error folder
}
Onet codes files validation
\enumerate{
\item Verify that the number of available Onet codes in the file at Country level is equal to the number of Onet codes available in the State level file
\item Verify that the sum of total supply crossed Onet codes at State level is equal to Country level
\item Verify that the sum of supply by occupation at Country level is equal than its State level counterpart. If test #3 fails, a file with all codes for which this test does not respect the constraint is saved in the error folder
\item Verify that the sum by Onet occupation at MSA level is equal or less than its State counterpart.If test #4 fails, a file with all codes for which this test does not respect the constraint is saved in the error folder
}
}
\examples{
supply_validation(2017,"South_Africa","user","password",20,4,"yes","10.0.5.30")
}
\keyword{supply}
